move buildError to Helper
