+++ packages/ts-predicate/docs/Helper.md
@@ -14,7 +14,7 @@ Return true in the following cases :
 
 Otherwise, return false.
 
-## normalizeErrorTree
+## NormalizeErrorTree
 
 ```ts
 normalizeErrorTree(error: Error): NormalizedError
@@ -34,13 +34,13 @@ Useful when using `TypeAssertion` functions to provide a response explaining wha
 
 `causes` is filled with the normalized errors of an `AggregateError`, or the normalized cause of an `Error` if it exists.
 
-## stringifyErrorTree
+## StringifyErrorTree
 
 ```ts
 stringifyErrorTree(error: Error | NormalizedError): string
 ```
 
-Convert an Error or a normalized error tree structure into a printable string.
+Convert an `Error` or `NormalizedError` into a printable string.
 
 An example of output:
 
@@ -50,17 +50,28 @@ The value is an object, but some properties are incorrect.
 ├─── The value has an extraneous property "LastName".
 ├─── The required property "firstName" is missing.
 ├─── The required property "lastName" is missing.
-├─── The property "birthDate" have an incorrect value.
+├─── The property "birthDate" has an incorrect value.
 │    └─── The value must be an instance of Date.
-└─── The property "hobbies" have an incorrect value.
+└─── The property "hobbies" has an incorrect value.
      └─── The value is an array, but its content is incorrect.
           └─── The value at index 1 is incorrect.
                └─── The value must be a string.
 ```
 
-Useful when using `TypeAssertion` functions to print what's wrong into the console.
+Useful when using `TypeAssertion` functions to print what's wrong in the console.
 
-## createCopiesWithMissingProperty
+## ToError
+
+```ts
+toError(reason: unknown): Error
+```
+
+if the parameter `reason` is an instance of `Error` it is returned as is,
+otherwise an instance of `UnknownError` is returned.
+
+`UnknownError` extends `Error` and has a property `reason` with the original value.
+
+## CreateCopiesWithMissingProperty
 
 ```ts
 createCopiesWithMissingProperty<T>(value: T): Array<Partial<T>>
@@ -68,7 +79,7 @@ createCopiesWithMissingProperty<T>(value: T): Array<Partial<T>>
 
 Helper for testing. It returns an array of shallow copies of the object, but for each one, a different property will be missing entirely.
 
-## createCopiesWithNullishProperty
+## CreateCopiesWithNullishProperty
 
 ```ts
 createCopiesWithNullishProperty<T>(value: T): Array<Nullable<T>>
@@ -76,7 +87,7 @@ createCopiesWithNullishProperty<T>(value: T): Array<Nullable<T>>
 
 Helper for testing. It returns an array of shallow copies of the object, but for each one, a different property will be replaced by `undefined`.
 
-## createCopiesWithInvalidProperty
+## CreateCopiesWithInvalidProperty
 
 ```ts
 createCopiesWithInvalidProperty<T>(value: T, mapping: Invalid<T>): Array<Invalid<T>>
