+++ packages/ts-predicate/src/Helper/_index.mts
@@ -4,3 +4,5 @@ export * from "./createCopiesWithNullishProperty.mjs";
 export * from "./isSimilar.mjs";
 export * from "./normalizeErrorTree.mjs";
 export * from "./stringifyErrorTree.mjs";
+export * from "./toError.mjs";
+export * from "./utils/UnknownError.mjs";
