+++ packages/ts-predicate/src/index.mts
@@ -1,6 +1,5 @@
 export * as Helper from "./Helper/_index.mjs";
 export * as TypeAssertion from "./TypeAssertion/_index.mjs";
-export * from "./TypeAssertion/utils/UnknownError.mjs";
 export * as TypeGuard from "./TypeGuard/_index.mjs";
 export * as TypeHint from "./TypeHint/_index.mjs";
 export * from "./types/_index.mjs";
